Implemented scoring system
The nesting algorithm now uses a scoring system to choose the relevant background for a given unnested genotype. Previously candidates where chosen based on a set of pre-defined criteria, then the candidate which occured closest to the unnested genotype was chosen as the background. Now each candidate is given a score based on how well it passes the inheritance checks and the candidate with the greatest score (detection time point breaks ties) is chosen as the background. Also changed the logging system.