fixed: Not specifying the output folder would cause the program to fail.
If the output folder was not specified by the user, the output folder was inferred form the `--input` flag. However, the output_folder variable was then a str object, which caused an AttributeError when testing whether the folder already existed. This has been fixed by converting the output_folder to a Path object more explicitly when cleaning up the program parameters.